+++ OpenRA.Mods.Common/Traits/Sound/ActorLostNotification.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	class ActorLostNotificationInfo : ITraitInfo
 	{
